pyproject.toml
[tool.poetry]
name = "sexify"
version = "1.0.0"
description = "Lossless. Effortless. 💿 Your Music, Uncompressed. Download Spotify tracks in FLAC via Tidal, Qobuz & Amazon Music."
authors = ["ManOfInfinity <ManOfInfinity@users.noreply.github.com>"]
readme = "README.md"
license = "MIT"
homepage = "https://github.com/ManOfInfinity/sexify"
repository = "https://github.com/ManOfInfinity/sexify"
keywords = ["spotify", "flac", "music", "downloader", "tidal", "qobuz", "amazon", "lossless"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Sound/Audio",
]
packages = [{include = "sexify"}]

[tool.poetry.dependencies]
python = "^3.9"
click = "^8.1.7"
requests = "^2.31.0"
mutagen = "^1.47.0"
pyyaml = "^6.0.1"
pillow = "^10.2.0"
rich = "^14.2.0"

[tool.poetry.scripts]
sexify = "sexify.cli:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
